refactor(Post): Fix fragment colocation
